feat: move ui logger into services
