+++ src/ui/pages/home/Home.tsx
@@ -1,6 +1,6 @@
 import { Button } from "../../components/common/button/Button";
 import type { FileRejection } from "@mantine/dropzone";
-import { loggerUi } from "../../utils/loggerUi";
+import { loggerUi } from "../../services/UiLoggingService.ts";
 import { Dropzone } from "../../components/common/dropzone/Dropzone.tsx";
 import { pushWarningNotification } from "../../services/NotificationService.ts";
 import { useFileManagement } from "../../hooks/useFileManagement.ts";
